pyproject.toml
[project]
name = "nonebot-plugin-parser"
version = "2.3.9"
description = "NoneBot2 链接分享解析 Alconna 版, 通用媒体卡片渲染(PIL 实现), 支持 B站/抖音/快手/微博/小红书/youtube/tiktok/twitter/acfun/nga"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ "name" = "fllesser", "email" = "fllessive@gmail.com" }]
keywords = [
  "acfun",
  "bilibili",
  "douyin",
  "kuaishou",
  "nga",
  "nonebot",
  "nonebot2",
  "tiktok",
  "twitter",
  "video",
  "weibo",
  "xiaohongshu",
  "youtube",
]

classifiers = [
  "Development Status :: 5 - Production/Stable",
  "Framework :: AsyncIO",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Communications :: Chat",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Multimedia :: Video",
]

dependencies = [
  "nonebot2>=2.4.3,<3.0.0",
  "pillow>=11.0.0",
  "tqdm>=4.67.1,<5.0.0",
  "aiofiles>=25.1.0",
  "httpx>=0.27.2,<1.0.0",
  "msgspec>=0.20.0,<1.0.0",
  "apilmoji[tqdm]>=0.2.4,<1.0.0",
  "beautifulsoup4>=4.12.0,<5.0.0",
  "curl_cffi>=0.13.0,<1.0.0,!=0.14.0",
  "bilibili-api-python>=17.4.1,<18.0.0",
  "nonebot-plugin-alconna>=0.59.4,<1.0.0",
  "nonebot-plugin-apscheduler>=0.5.0,<1.0.0",
  "nonebot-plugin-localstore>=0.7.4,<1.0.0",
  "nonebot-plugin-uninfo>=0.10.1,<1.0.0",
]

[project.urls]
IssueTracker = "https://github.com/fllesser/nonebot-plugin-parser/issues"
Release = "https://github.com/fllesser/nonebot-plugin-parser/releases"
Repository = "https://github.com/fllesser/nonebot-plugin-parser"

[project.optional-dependencies]
htmlkit = ["nonebot-plugin-htmlkit>=0.1.0rc4"]
htmlrender = ["nonebot-plugin-htmlrender>=0.6.7"]
ytdlp = ["yt-dlp[default]>=2026.1.29"]
emosvg = ["emosvg>=0.1.6"]
all = [
  "nonebot-plugin-htmlkit>=0.1.0rc4",
  "nonebot-plugin-htmlrender>=0.6.7",
  "yt-dlp[default]>=2026.1.29",
  "emosvg>=0.1.6",
]

[dependency-groups]
dev = [
  "nonebot2[fastapi]>=2.4.3,<3.0.0",
  "nonebot-adapter-onebot>=2.4.6",
  "ruff>=0.14.14,<1.0.0",
  { include-group = "extras" },
  { include-group = "test" },
]
extras = [
  "emosvg>=0.1.6",
  "nonebot-plugin-htmlkit>=0.1.0rc4",
  "nonebot-plugin-htmlrender>=0.6.7",
  "yt-dlp[default]>=2026.1.29",
  "types-yt-dlp>=2026.1.29.20260131",
]
test = [
  "nonebug>=0.4.3,<1.0.0",
  "poethepoet>=0.40.0",
  "pytest-asyncio>=1.3.0,<1.4.0",
  "pytest-cov>=7.0.0",
  "pytest-xdist>=3.8.0,<4.0.0",
  "respx>=0.22.0",
]
pydantic-v1 = ["pydantic<2.0.0"]
pydantic-v2 = ["pydantic>=2.0.0"]
telegram = ["nonebot-adapter-telegram>=0.1.0b20"]

[build-system]
requires = ["uv_build>=0.9.0,<0.10.0"]
build-backend = "uv_build"

[tool.uv]
conflicts = [
  [
    { group = "pydantic-v1" },
    { group = "pydantic-v2" },
    { group = "telegram" },
  ],
]

[tool.uv.sources]
nonebug = { git = "https://github.com/nonebot/nonebug" }

[tool.bumpversion]
tag = true
commit = true
current_version = "2.3.9"
message = "release: bump vesion from {current_version} to {new_version}"

[[tool.bumpversion.files]]
filename = "uv.lock"
search = "name = \"nonebot-plugin-parser\"\nversion = \"{current_version}\""
replace = "name = \"nonebot-plugin-parser\"\nversion = \"{new_version}\""

[tool.coverage.report]
show_missing = true
exclude_lines = [
  "raise NotImplementedError",
  "if TYPE_CHECKING:",
  "@overload",
  "except",
  "raise",
]

[tool.nonebot]
adapters = [
  { "name" = "Onebot V11", module_name = "nonebot.adapters.onebot.v11" },
  # { name = "Telegram", module_name = "nonebot.adapters.telegram" },
]
plugins = ["nonebot_plugin_parser"]

[tool.poe.tasks]
test-other = "pytest tests/others --cov=src --cov-report=xml --junitxml=junit.xml -n auto"
test-parser = "pytest tests/parsers --cov=src --cov-report=xml --junitxml=junit.xml -n auto"
test-render = "pytest tests/renders --cov=src --cov-report=xml --junitxml=junit.xml"
bump = "bump-my-version bump"
show-bump = "bump-my-version show-bump"

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
  { root = "./tests", extraPaths = [
    "./src",
  ] },
  { root = "./src" },
]
typeCheckingMode = "standard"
disableBytesTypePromotions = true

[tool.pytest]
addopts = [
  "--import-mode=prepend", # 导入模式
  "--strict-markers",      # 严格标记模式
  "--tb=short",            # 简短的错误回溯
  "-ra",                   # 显示所有测试结果摘要
  "-s",                    # 显示打印信息
  "-v",                    # 详细输出
]
pythonpath = ["src"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"

[tool.ruff]
line-length = 120

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
  "F",     # Pyflakes
  "W",     # pycodestyle warnings
  "E",     # pycodestyle errors
  "I",     # isort
  "UP",    # pyupgrade
  "Q",     # flake8-quotes
  "PYI",   # flake8-pyi
  "T10",   # flake8-debugger
  "T20",   # flake8-print
  "C4",    # flake8-comprehensions
  "PT",    # flake8-pytest-style
  "TID",   # flake8-tidy-imports
  "ASYNC", # flake8-async
  "RUF",   # Ruff-specific rules
]
ignore = [
  "E402",   # module-import-not-at-top-of-file
  "UP037",  # quoted-annotation
  "RUF001", # ambiguous-unicode-character-string
  "RUF002", # ambiguous-unicode-character-docstring
  "RUF003", # ambiguous-unicode-character-comment
  "W191",   # indentation contains tabs
  "TID252", # relative-import
]

[tool.ruff.lint.isort]
length-sort = true
known-first-party = ["nonebot_plugin_parser", "tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true
